refactor: 將重複宣告的介面抽成常用介面
調整項目:
  1. 將重複宣告的介面抽成常用介面。
