+++ src/app/views/shop-management/good-management/good-management.component.ts
@@ -7,13 +7,14 @@ import { BreadcrumbService } from 'src/app/services/breadcrumb-service/breadcrum
 import { AppEnvironmentService } from 'src/app/services/app-environment-service/app-environment.service';
 import { RequestService } from 'src/app/services/request-service/request.service';
 import Modal from 'bootstrap/js/dist/modal';
-import { AddGood, AsGoodType, Good, GoodManagementStatuses, Modals, PageInformation, SelectedImage } from './good-management';
+import { AddGood, AsGoodType, Good, GoodManagementStatuses } from './good-management';
 import { ApiServiceTypes } from 'src/app/enums/api-service-types';
 import { HttpErrorResponse } from '@angular/common/http';
 import { v4 as uuidv4 } from 'uuid';
 import { BaseResponse } from 'src/app/commons/abstracts/http-client';
 import { SingleFileUploadResponse } from 'src/app/commons/abstracts/file-storage-service';
 import { CKEditorModule } from 'ckeditor4-angular';
+import { Modals, PageInformation, SelectedImage } from 'src/app/commons/abstracts/common';
 
 @Component({
   selector: 'app-good-management',
